requirements.txt
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
seaborn==0.13.2
statsmodels==0.14.2
scikit-learn==1.5.2
plotly==5.23.0
streamlit==1.37.1
python-pptx==0.6.23